[Chores] StarUserDefaultsManagerTester 파일 이동
